projects/energy-disaggregation/requirements.txt
# Runtime dependencies — keep this file up to date with `pip freeze > requirements.txt`
# after installing packages into your .venv.

python-dotenv>=1.0.0
numpy>=1.26.0
